requirements.txt
Flask
requests-oauthlib
Flask-Dance
python-dotenv
Flask-Login
Flask-WTF
Flask-Cors
WTForms
psycopg2-binary
Werkzeug
gunicorn
google-genai
Pillow
cloudinary
email-validator
reportlab
html5lib
six
flask-mail
itsdangerous
python-bidi==0.4.2
xhtml2pdf==0.2.16
groq